@@ -1,32 +1,32 @@
 [build-system]
-build-backend = "setuptools.build_meta"
+build-backend = "rapids_build_backend.build"
 requires = [
+    "rapids-build-backend>=0.3.0,<0.4.0.dev0",
     "setuptools>=64.0.0",
-    "tomli  ; python_version < '3.11'",
-]
+] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
 
 [project]
 name = "distributed-ucxx"
-version = "0.36.00"
+dynamic = ["version"]
 description = "UCX communication module for Dask Distributed"
 readme = { file = "README.md", content-type = "text/markdown" }
 authors = [
     { name = "NVIDIA Corporation" },
 ]
-license = { text = "Apache-2.0" }
-requires-python = ">=3.8"
+license = { text = "BSD-3-Clause" }
+requires-python = ">=3.9"
 dependencies = [
-    "dask >=2023.9.2",
-    "distributed >=2023.9.2",
-    "numba >=0.54",
-]
+    "numba>=0.57.1",
+    "rapids-dask-dependency==24.10.*,>=0.0.0a0",
+    "ucxx==0.40.*,>=0.0.0a0",
+] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
 classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: Apache Software License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
 ]
 
 [project.entry-points."distributed.comm.backends"]
@@ -39,9 +39,12 @@ docs = [
     "sphinx-rtd-theme>=0.5.1",
 ]
 test = [
-    "pytest",
-    "numpy",
-]
+    "cudf==24.10.*,>=0.0.0a0",
+    "cupy-cuda11x>=12.0.0",
+    "numpy>=1.23,<2.0a0",
+    "pytest-rerunfailures",
+    "pytest==7.*",
+] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
 
 [project.urls]
 Homepage = "https://github.com/rapidsai/ucxx"
@@ -111,10 +114,18 @@ exclude = [
     "tests.*",
 ]
 
-[tool.versioneer]
-VCS = "git"
-style = "pep440"
-versionfile_source = "distributed_ucxx/_version.py"
-versionfile_build = "distributed_ucxx/_version.py"
-tag_prefix = "v"
-parentdir_prefix = "distributed_ucxx-"
+[tool.rapids-build-backend]
+build-backend = "setuptools.build_meta"
+dependencies-file = "../../dependencies.yaml"
+
+[tool.setuptools.package-data]
+"distributed_ucxx" = ["VERSION"]
+
+[tool.setuptools.dynamic]
+version = {file = "distributed_ucxx/VERSION"}
+
+[tool.pytest.ini_options]
+markers = [
+    "ignore_alive_references",
+    "slow",
+]
