@@ -6,6 +6,7 @@
 #include <ucxx/delayed_submission.h>
 #include <ucxx/internal/request_am.h>
 #include <ucxx/request_am.h>
+#include <ucxx/typedefs.h>
 
 namespace ucxx {
 
@@ -14,7 +15,8 @@ namespace internal {
 RecvAmMessage::RecvAmMessage(internal::AmData* amData,
                              ucp_ep_h ep,
                              std::shared_ptr<RequestAm> request,
-                             std::shared_ptr<Buffer> buffer)
+                             std::shared_ptr<Buffer> buffer,
+                             AmReceiverCallbackType receiverCallback)
   : _amData(amData), _ep(ep), _request(request)
 {
   std::visit(data::dispatch{
@@ -22,6 +24,12 @@ RecvAmMessage::RecvAmMessage(internal::AmData* amData,
                [](auto) { throw std::runtime_error("Unreachable"); },
              },
              _request->_requestData);
+
+  if (receiverCallback) {
+    _request->_callback = [this, receiverCallback](ucs_status_t, std::shared_ptr<void>) {
+      receiverCallback(_request);
+    };
+  }
 }
 
 void RecvAmMessage::setUcpRequest(void* request) { _request->_request = request; }
